#27964: Provide protected access to read and write step analysis properties
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@21273 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Controller/src/main/java/org/gusdb/wdk/controller/action/analysis/AbstractStepAnalysisIdAction.java
@@ -45,8 +45,8 @@ public abstract class AbstractStepAnalysisIdAction extends WdkAction {
     return PARAMS;
   }
   
-  protected final int getAnalysisId() {
-    return getParams().getIntValue(StepAnalysisContext.ANALYSIS_ID_KEY);
+  protected final long getAnalysisId() {
+    return getParams().getLongValue(StepAnalysisContext.ANALYSIS_ID_KEY);
   }
   
   protected final StepAnalysisFactory getAnalysisMgr() {
